Refactor: Update button shapes and remove card shadows
This commit updates the button shapes in the `ToDoListScreen` to use `ButtonShapes` with a `RoundedCornerShape` for the pressed state.

Additionally, the shadow effect previously applied to `ElevatedCard` components using `CardDefaults.elevatedShape` has been removed.

Keyboard actions have also been added to the description TextField to allow saving/updating the to-do item upon pressing the "Done" action on the keyboard.