Adjust description color based on todo status
This commit modifies the `TodoItem` composable to change the text color of the todo description based on its completion status.

If a todo is marked as done (`isDone` is true), the description text will now appear in gray. Otherwise, it will use the default unspecified color. This visual cue helps users quickly differentiate between completed and pending tasks.